requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.40.0
pillow>=10.0.0

# Data generation
MetaTrader5>=5.0.0
matplotlib>=3.8.0
pandas>=2.0.0
numpy>=1.24.0

# Acceleration (optional but recommended)
accelerate>=0.27.0
